v3.0.0a12-4
Codebase changes:
- Created multiple controllers and helper functions for a better maintainability
