+++ app/insertVersion.js
@@ -1,3 +1,3 @@
-appVersion = "3.0.0a12";
+appVersion = "3.0.0a12-4";
 
 $(".version").text(appVersion);
