+++ app/utils/misc.js
@@ -1,0 +1,3 @@
+export function delay(ms) {
+    return new Promise((resolve) => setTimeout(resolve, ms));
+  }
